# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gelasha Market Place API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SignupRequest {
  username: string;
  email: string;
  password: string;
  siteName: string;
  siteAddress: string;
  transportAddress?: string;
  gstNumber?: string;
  mobile1: string;
  mobile2?: string;
  appUsage?: string;
  fuelType?: string;
}

export interface SigninRequest {
  email: string;
  password: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  siteName?: string;
  siteAddress?: string;
  transportAddress?: string;
  gstNumber?: string;
  mobile1?: string;
  mobile2?: string;
  appUsage?: string;
  fuelType?: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  price: number;
  originalPrice?: number;
  category: string;
  imageUrl: string;
  images: string[];
  rating: number;
  reviewCount: number;
  inStock: boolean;
  brand: string;
  discount?: number;
}

export interface Order {
  id: string;
  userId: string;
  productId: string;
  product?: Product;
  quantity: number;
  totalPrice: number;
  status: string;
  createdAt: string;
}

export interface CreateOrderRequest {
  userId: string;
  productId: string;
  quantity: number;
}

export interface ToggleLikeRequest {
  userId: string;
  productId: string;
}

export interface LikeResponse {
  liked: boolean;
}

export type GetMeParams = {
  userId: string;
};

export type GetProductsParams = {
  category?: string;
  search?: string;
};

export type GetOrdersParams = {
  userId: string;
};

export type GetLikesParams = {
  userId: string;
};