# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gelasha Market Place API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: products
    description: Product operations
  - name: orders
    description: Order operations
  - name: likes
    description: Like operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/signin:
    post:
      operationId: signin
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninRequest"
      responses:
        "200":
          description: Signed in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /likes:
    get:
      operationId: getLikes
      tags: [likes]
      summary: Get liked products for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of liked products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: toggleLike
      tags: [likes]
      summary: Toggle like on a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleLikeRequest"
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"

  /categories:
    get:
      operationId: getCategories
      tags: [products]
      summary: Get all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SignupRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        siteName:
          type: string
        siteAddress:
          type: string
        transportAddress:
          type: string
        gstNumber:
          type: string
        mobile1:
          type: string
        mobile2:
          type: string
        appUsage:
          type: string
        fuelType:
          type: string
      required:
        - username
        - email
        - password
        - siteName
        - siteAddress
        - mobile1

    SigninRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        siteName:
          type: string
        siteAddress:
          type: string
        transportAddress:
          type: string
        gstNumber:
          type: string
        mobile1:
          type: string
        mobile2:
          type: string
        appUsage:
          type: string
        fuelType:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - role
        - createdAt

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        brand:
          type: string
        discount:
          type: integer
      required:
        - id
        - name
        - description
        - price
        - category
        - imageUrl
        - images
        - rating
        - reviewCount
        - inStock
        - brand

    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - productId
        - quantity
        - totalPrice
        - status
        - createdAt

    CreateOrderRequest:
      type: object
      properties:
        userId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
      required:
        - userId
        - productId
        - quantity

    ToggleLikeRequest:
      type: object
      properties:
        userId:
          type: string
        productId:
          type: string
      required:
        - userId
        - productId

    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
      required:
        - liked