# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gelasha Market Place API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderRequest {
  userId: string;
  productId: string;
  quantity: number;
}