# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gelasha Market Place API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface Order {
  id: string;
  userId: string;
  productId: string;
  product?: Product;
  quantity: number;
  totalPrice: number;
  status: string;
  createdAt: string;
}