# lib/api-zod/src/generated/types/signinRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gelasha Market Place API
 * OpenAPI spec version: 0.1.0
 */

export interface SigninRequest {
  email: string;
  password: string;
}