# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gelasha Market Place API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  username: string;
  email: string;
  siteName?: string;
  siteAddress?: string;
  transportAddress?: string;
  gstNumber?: string;
  mobile1?: string;
  mobile2?: string;
  appUsage?: string;
  fuelType?: string;
  role: string;
  createdAt: string;
}